refactor: history inferface
