+++ src/logic/history/task_history.py
@@ -3,7 +3,6 @@
     the interface for the history of completed tasks.
 
     Classes:
-        HistoryInterface(ABC): Abstract Base Class
         HistorySingleton(HistoryInterface): Interface of the history of completed tasks,
         used to create a singleton instance of the class.
 
@@ -11,30 +10,9 @@
         add_completed_task(self, project: Item) -> None: Adds a completed task to the list
         tasks_completed(self) -> List[Item]: List of completed tasks
 """
-
-from abc import ABC, abstractmethod
 from typing import List
 from src.logic.items.item_interface import IItem
-
-class HistoryInterface(ABC):
-    """ This class will be used to represent the interface history of completed tasks
-
-    Args:
-        ABC (): Abstract Base Class
-    """
-    @abstractmethod
-    def add_completed_task(self, project: IItem) -> None:
-        """ This method will be used to add a completed task to the list
-
-        Args:
-            task_name (Item): Adds a completed task to the list
-        """
-
-    @abstractmethod
-    def tasks_completed(self) -> List[IItem]:
-        """ This method will be used to return the list of completed tasks
-        """
-
+from src.logic.history.history_interface import HistoryInterface
 class HistorySingleton(HistoryInterface):
     """ This class will be used to represent the history of completed tasks
